test(browser): Update the test app to include soft navigation action
