+++ dev-packages/e2e-tests/test-applications/default-browser/public/index.html
@@ -10,6 +10,14 @@
 
     <input type="button" value="Capture Exception" id="exception-button" />
 
+    <div id="navigation">
+
+      <a id="navigation-link" href="#navigation-target">Navigation Link</a>
+
+      <div id="navigation-target">Navigated Element</div>
+      
+    </div>
+
     <!-- The script tags for the bundled JavaScript files will be injected here by HtmlWebpackPlugin in build.mjs-->
   </body>
 </html>
